feat: #526 organize bundler extension APIs
Put all bundler stuff in one folder so that we can remove or migrate it to the
bundler project later.
