+++ packages/liferay-js-toolkit-core/src/bundler/plugin.ts
