+++ packages/liferay-js-toolkit-core/src/index.ts
@@ -21,14 +21,14 @@ export type {
 	BundlerLoaderEntryPoint,
 	BundlerLoaderContext,
 	BundlerLoaderReturn,
-} from './api/loaders';
+} from './bundler/loader';
 export type {
 	BabelIpcObject,
 	BundlerPluginEntryPoint,
 	BundlerPluginParams,
 	BundlerCopyPluginState,
 	BundlerTransformPluginState,
-} from './api/plugins';
+} from './bundler/plugin';
 export * as babelIpc from './babel-ipc';
 export * from './babel-util'; // TODO: not sure if this is needed in bundler 3
 export {default as PkgDesc} from './pkg-desc';
